pyproject.toml
[project]
name = "digit-recogniser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "streamlit-drawable-canvas>=0.9.3",
    "torch>=2.7.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.22.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
